ProfileForms component added
