+++ src/components/profile-forms/profile-forms.module.css
@@ -1,0 +1,22 @@
+.userData {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  align-items: center;
+  grid-area: userData;
+}
+
+.paragraph {
+  grid-area: paragraph;
+  opacity: 0.4;
+}
+
+.buttonsContainer {
+  grid-area: buttonsContainer;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.container input {
+  color: #8585ad;
+}
